refactor(coordinator): extract polish/translate orchestration into coordinator/polish_flow.rs — behavior unchanged, 446 tests green
Pure move of the StreamingPolishOutcome enum + polish_or_passthrough_streaming /
polish_or_passthrough / polish_text / translate_text / polish_and_translate_or_passthrough
and the polish-translate marker consts/helpers, from coordinator.rs into a new sibling
submodule. Made pub(super); reached from the parent and from dictation.rs via
'use polish_flow::*;'. No logic/signature/public-API changes; coordinator.rs 5285→4934 lines.

cargo check: 0 errors, 87 warnings (unchanged) · clippy: 0 errors
cargo test --lib: 446 passed · tsc / npm run build: green
